test: new vote
