+++ tests/transactions/serializers/test_vote.py
@@ -1,6 +1,6 @@
 from solar_crypto.transactions.serializer import Serializer
 
 
-def test_serializer(transaction_type_3):
-    result = Serializer(transaction_type_3).serialize(False, True)
-    assert result == transaction_type_3["serialized"]
+def test_serializer(transaction_type_vote):
+    result = Serializer(transaction_type_vote).serialize(False, True)
+    assert result == transaction_type_vote["serialized"]
